+++ backend/accounts/account.go
@@ -86,9 +86,9 @@ type Interface interface {
 	Transactions() (OrderedTransactions, error)
 	// Must enforce that initial sync is done before returning.
 	Balance() (*Balance, error)
-	// SendTx signs and sends the active tx proposal, set by TxProposal. Errors if none
-	// available.
-	SendTx(txNote string) error
+	// SendTx signs and sends the active tx proposal, set by TxProposal, and returns its
+	// id. Errors if none available.
+	SendTx(txNote string) (string, error)
 	FeeTargets() ([]FeeTarget, FeeTargetCode)
 	TxProposal(*TxProposalArgs) (coin.Amount, coin.Amount, coin.Amount, error)
 	// GetUnusedReceiveAddresses gets a list of list of receive addresses. The result can be one